Add files via upload
We did this project on the chatbot google dialogflow. We know that in our daily life we have some real time examples that are amazon lex. Rasa and microsoft
And then we are using this dialog flow web site and then we are able to see the intents and entities in our left corner. But we are using only intents for this chatbot. We 
Have to make some intent with our name or our source name and then we have use intent and make that training phrases what we have to use. Then we have to give
Certain responses and then we have to save it with the name of the same intent. Then we will get the answer what we have given in the responses defaulty when we give
 the training phrases as our inputs. Then we have to integrate that dialogflow in the chatbot of the talegram by going to the telegram and registering with the fatherbot. 
Then we are able to see our default responses in telegram.